Add files via upload
I found this project to be much less challenging than the most recent projects. It was really simple for me to create the shapes, get them to move, and have text appear. I may have done it in a less than ideal way from a backend perspective, but the game functions perfectly fine!